add filter and sort component integration
