+++ apps/api/src/routes/movies.js
@@ -4,4 +4,8 @@ module.exports = app => {
 	app.route('/movies').get(controller.listMovies);
 
 	app.route('/movies/:id').get(controller.getMovie);
+
+	app.route('/genres').get(controller.getGenresList);
+
+	app.route('/sort/:sort/:genreId').get(controller.getSort);
 };
