+++ apps/themovie/package.json
@@ -8,6 +8,7 @@
 		"@testing-library/jest-dom": "^5.11.6",
 		"@testing-library/react": "^11.2.2",
 		"@testing-library/user-event": "^12.2.2",
+		"axios": "^0.21.0",
 		"history": "^5.0.0",
 		"react": "^17.0.1",
 		"react-dom": "^17.0.1",
