--- apps/themovie/src/components/MovieCard/renderMovieCard.js
@@ -1,5 +1,0 @@
-import MovieCard from './MovieCard'
-
-
-
-export const renderMoviesCard = ( movie, index ) => ( <MovieCard key={`movie-${index}`} {...movie} /> )
