+++ apps/themovie/src/pages/Home/index.js
@@ -1,7 +1,9 @@
 import React from 'react';
 
 const Home = () => {
-	return <div></div>;
+	return (
+		<div></div>
+	)
 };
 
 export default Home;
